src/main/java/shoppingmall/ankim/domain/payment/service/PaymentService.java
package shoppingmall.ankim.domain.payment.service;

import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;
import shoppingmall.ankim.domain.order.entity.Order;
import shoppingmall.ankim.domain.order.exception.OrderNotFoundException;
import shoppingmall.ankim.domain.order.repository.OrderRepository;
import shoppingmall.ankim.domain.payment.dto.PaymentResponse;
import shoppingmall.ankim.domain.payment.entity.Payment;
import shoppingmall.ankim.domain.payment.repository.PaymentRepository;
import shoppingmall.ankim.domain.payment.service.request.PaymentCreateServiceRequest;
import shoppingmall.ankim.global.config.TossPaymentConfig;
import shoppingmall.ankim.global.exception.ErrorCode;

import java.util.Optional;

import static shoppingmall.ankim.global.exception.ErrorCode.ORDER_NOT_FOUND;

@Service
@Transactional
@RequiredArgsConstructor
public class PaymentService {
    private final PaymentRepository paymentRepository;
    private final TossPaymentConfig tossPaymentConfig;
    private final OrderRepository orderRepository;

    public PaymentResponse requestTossPayment(PaymentCreateServiceRequest request) {
        // Order 조회 (fetch join으로 Member 로딩)
        Order order = orderRepository.findByOrderCodeWithMember(request.getOrderCode())
                .orElseThrow(() -> new OrderNotFoundException(ORDER_NOT_FOUND));

        // Payment 생성 & 저장
        Payment payment = paymentRepository.save(Payment.create(order, request.getPayType(), request.getAmount()));

        // PaymentResponse 변환
        PaymentResponse response = PaymentResponse.of(payment);
        response.setSuccessUrl(request.getYourSuccessUrl() == null ? tossPaymentConfig.getSuccessUrl() : request.getYourSuccessUrl());
        response.setFailUrl(request.getYourFailUrl() == null ? tossPaymentConfig.getFailUrl() : request.getYourFailUrl());

        return response;
    }


}
